refactor: move out errors from lib to a standalone module
